+++ benches/bench_10d_gaussian_profile.rs
@@ -44,6 +44,7 @@ fn main() {
         .map(|c| TermSpec::Cr {
             col: c,
             k: k_vec[c],
+            pc: None,
         })
         .collect();